acore(Mfuzz)	R Documentation

Extraction of alpha cores for soft clusters

Description

This function extracts genes forming the alpha-cores of soft clusters

Usage

acore(eset,cl,min.acore=0.5,anno=FALSE,html=FALSE)

Arguments

eset 	 object of the class“exprSet”

cl 	 An object of class flcust as produced by mfuzz.

min.acore 	 minimum membership values of a gene to belong to the core

anno 	 If anno=TRUE annotations are includecd in the output list.

html 	 If html=TRUE a HTML file (‘alpha.cores.html’) with alpha cores is produced.

Value

The function produces an list of alpha cores including genes and their membership values for the corresponding cluster.

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

Examples


data(yeast) # data set includes 17 measurements
yeastF <- filter.NA(yeast) 
yeastF <- fill.NA(yeastF)
yeastF <- filter.NA(yeast)
yeastF <- fill.NA(yeastF)
yeastF <- standardise(yeastF)

# Soft clustering and visualisation
cl <- mfuzz(yeastF,c=20,m=1.25)
acore.list <- acore(yeastF,cl=cl,min.acore=0.7,anno=TRUE,html=TRUE)

[Package Mfuzz version 0.0.1 Index]