cselection(Mfuzz)	R Documentation

Repeated soft clustering for detection of empty clusters

Description

This function repeatly performs soft clustering for a range of cluster numbers and reports the number of empty clusters detected.

Usage

cselection(eset,m,crange=seq(4,32,4),repeats=5,visu=TRUE,...)

Arguments

eset 	 object of class “exprSet”

m 	 value of fuzzy c-means parameter m

crange 	 range of number of clusters c

repeats 	 number of repeated clusterings

visu 	 if visu=TRUE plot of number of empty clusters is produced.

... 	 additional arguments for underlying mfuzz

Details

A soft cluster is considered as empty, if none of the genes has a correponding membership value larger than 0.5

Value

The function generates the matrix of the number of empty clusters detected.

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

References

M.E. Futschik and B. Charlisle, Noise robust clustering of gene expression time-course data, Journal of Bioinformatics and Computational Biology, 3 (4), 2005

Examples


data(yeast)
# Data pre-processing
yeastF <- filter.NA(yeast)
yeastF <- fill.NA(yeastF)
yeastF <- standardise(yeastF)

#### parameter selection
yeastFR <- randomise(yeastF)
cl <- mfuzz(yeastFR,c=16,m=1.25)
mfuzz.plot(yeastFR,cl=cl,mfrow=c(4,5))

tmp  <- partcoef(yeastFR)
F <- tmp[[1]];F.n <- tmp[[2]];F.min <- tmp[[3]]
F > F.min

[Package Mfuzz version 0.0.2 Index]