filter.NA(Mfuzz)	R Documentation

Filtering of genes based on number of non-available expression values.

Description

This function can be used to exclude genes with a large number of expression values not available.

Usage

filter.NA(eset,thres=0.25)

Arguments

eset 	 object of the class“exprSet”

thres 	 threshold for excluding genes. If the percentage of NA is larger than thres, the corresponding gene will be excluded.

Value

The function produces an object of the exprSet class. It is the same as the input eset object, except for the genes excluded.

Note

If eset has an annotation slot, the elements corresponding to the excluded genes will also be deleted.

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

Examples


data(yeast) # data set includes 17 measurements
yeastF <- filter.NA(yeast) # genes are excluded if more than 4 measurements are missing

[Package Mfuzz version 0.0.1 Index]