filter.std(Mfuzz)	R Documentation

Filtering of genes based on their standard deviation.

Description

This function can be used to exclude genes with low standard deviation.

Usage

filter.std(eset,min.std,visu=TRUE)

Arguments

eset 	 object of the class“exprSet”

min.std 	 threshold for minimum standard deviation. If the standard deviation of a gene's expresssion is smaller than min.std the corresponding gene will be excluded.

visu 	 If visu is set to TRUE, the ordered standard deviations of genes' expression values will be plotted.

Value

The function produces an object of the exprSet class. It is the same as the input eset object, except for the genes excluded.

Note

As soft clustering is noise robust, pre-filtering can usually be avoided. However, if the number of genes with small expression changes is large, such pre-filtering may be necessary to reduce noise. Note also:If eset has an annotation slot, the elements corresponding to the excluded genes will also be deleted.

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

Examples


data(yeast) # data set includes 17 measurements
yeastF <- filter.NA(yeast) # filtering of genes based on missing values 
yeastF <- filter.std(yeastF,min.std=0.3) # filtering of genes based on standard deviation 

[Package Mfuzz version 0.0.1 Index]