kmeans2(Mfuzz)	R Documentation

K-means clustering for gene expression data

Description

This function is a wrapper function for kmeans of the e1071 package. It performs soft clustering of genes based on their exprssion values using the k-means algorithm.

Usage

kmeans2(eset,k,iter.max=100)

Arguments

eset 	 object of the class “exprSet”

k 	 number of clusters

iter.max 	 maximal number of iterations

Value

An list of clustering components (see kmeans) .

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

See Also

kmeans

Examples


data(yeast)
# Data pre-processing
yeastF <- filter.NA(yeast)
yeastF <- fill.NA(yeastF)
yeastF <- standardise(yeastF)

# K-means clustering and visualisation
kl <- kmeans2(yeastF,k=20)
kmeans2.plot(yeastF,kl=kl,mfrow=c(2,2))


[Package Mfuzz version 0.0.2 Index]