mfuzz(Mfuzz)	R Documentation

Function for soft clustering based on fuzzy c-means.

Description

This function is a wrapper function for cmeans of the e1071 package. It performs soft clustering of genes based on their exprssion values using the fuzzy c-means algorithm.

Usage

mfuzz(eset,centers,m,...)

Arguments

eset 	 object of the class “exprSet”

centers 	 number of clusters

m 	 fuzzification parameter

... 	 additional parameters for cmeans

Value

An object of class flcust (see cmeans) .

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

See Also

cmeans

Examples


data(yeast)
# Data pre-processing
yeastF <- filter.NA(yeast)
yeastF <- fill.NA(yeastF)
yeastF <- standardise(yeastF)

# Soft clustering and visualisation
cl <- mfuzz(yeastF,c=20,m=1.25)
mfuzz.plot(yeastF,cl=cl,mfrow=c(2,2))


[Package Mfuzz version 0.1.0 Index]