mfuzz.plot(Mfuzz)	R Documentation

Plotting results for soft clustering

Description

This function visualises the clusters produced by mfuzz.

Usage

mfuzz.plot(eset,cl,mfrow=c(1,1),colo,min.mem=0)

Arguments

eset 	 object of the class“exprSet”

cl 	 object of class “flclust”

mfrow

colo 	 color palette to be used for plotting. If the color argument remains empty, the default palette is used.

min.mem 	 Genes with membership values below min.mem will not be displayed.

Value

The function generates plots where the membership of genes is color-encoded.

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

Examples


data(yeast)
# Data pre-processing
yeastF <- filter.NA(yeast)
yeastF <- fill.NA(yeastF)
yeastF <- standardise(yeastF)

# Soft clustering and visualisation
cl <- mfuzz(yeastF,c=20,m=1.25)
mfuzz.plot(yeastF,cl=cl,mfrow=c(2,2))
mfuzz.plot(yeastF,cl=cl,mfrow=c(3,3),min.mem=0.5)
mfuzz.plot(yeastF,cl=cl,mfrow=c(3,3),min.mem=0.7)

[Package Mfuzz version 0.0.2 Index]