overlap(Mfuzz)	R Documentation

Calculation of the overlap of soft clusters

Description

This function calculates the overlap of clusters produced by mfuzz.

Usage

overlap(cl)

Arguments

cl 	 object of class “flclust”

Value

The function generates a matrix of the normalised overlap of soft clusters. For the definiton of the overlap, see the vignette of the package.

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

Examples


data(yeast)
# Data pre-processing
yeastF <- filter.NA(yeast)
yeastF <- fill.NA(yeastF)
yeastF <- standardise(yeastF)

# Soft clustering and visualisation
cl <- mfuzz(yeastF,c=20,m=1.25)
mfuzz.plot(yeastF,cl=cl,mfrow=c(3,3))

# Calculation of cluster overlap and visualisation 
O <- overlap(cl)
X11()
Ptmp <- overlap.plot(cl,over=O,thres=0.05)

[Package Mfuzz version 0.0.2 Index]