overlap.plot(Mfuzz)	R Documentation

Visualisation of cluster overlap and global clustering structure

Description

This function visualises the cluster overlap produced by overlap.

Usage

overlap.plot(cl,overlap,thres=0.1,scale=TRUE,magni=30,P=NULL)

Arguments

cl 	 object of class “flclust”

overlap

thres 	 threshold for visualisation. Cluster overlaps below the threshold will not be visualised.

scale 	 Scale parameter for principal component analysis by prcomp

magni 	 Factor for increase the line width for cluster overlap.

P 	 Project/Rotation matrix produced by principal component analysis.

Value

A plot is genererated based on a prinicpal component analysis of the cluster centers. The overlap is visualised by lines with variable width determined by the strength of the overlap. Additonally, the matrix of principal components is generated. This matrix can be re-used for other projections to compare the overlap and global cluster structure of different clusterings.

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

See Also

prcomp

Examples


data(yeast)
# Data pre-processing
yeastF <- filter.NA(yeast)
yeastF <- fill.NA(yeastF)
yeastF <- standardise(yeastF)

# Soft clustering
cl <- mfuzz(yeastF,c=20,m=1.25)
X11();mfuzz.plot(yeastF,cl=cl,mfrow=c(4,5))
O <- overlap(cl)
X11();Ptmp <- overlap.plot(cl,over=O,thres=0.05)

# Alternative clustering
cl <- mfuzz(yeastF,c=10,m=1.25)
X11();mfuzz.plot(yeastF,cl=cl,mfrow=c(3,4))
O <- overlap(cl)
X11();overlap.plot(cl,over=O,P=Ptmp,thres=0.05)

[Package Mfuzz version 0.0.2 Index]