partcoef(Mfuzz)	R Documentation

Calculation of the partition coefficient matrix for soft clustering

Description

This function calculates partition coefficient for

Usage

partcoef(eset,crange=seq(4,32,4),mrange=seq(1.05,2,0.1),...)

Arguments

eset 	 object of class “exprSet”

crange 	 range of number of clusters c

mrange 	 range of clustering paramter m

... 	 additional arguments for underlying mfuzz

Details

Introduced by Bezdek (1981), the partition coefficient F is defined as the sum of squares of values of the partition matrix devided by the number of values. It is maximal if the partition is hard and reaches a minimum for U=1/c when every object is equally assigned to every cluster.

It is well-known that the partition coefficient tends to decrease monotonically with increasing n. To reduce this tendency we defined a normalized partition coefficient where the particion where is the partition coefficient derived from uniform partition (Futschik and Kasabov,2002).

Value

The function generates the matrix of partition coefficients for a range of c and m values. It also produces a matrix of normalised partition matrix as well as a matrix with matrix with partition coefficients for uniform partitions.

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

References

J.C.Bezdek, Pattern recognition with fuzzy objective function algorithms, Plenum, 1981
M.E. Futschik and N.K. Kasabov. Fuzzy clustering of gene expression data, Proceedings of World Congress of Computational Intelligence WCCI 2002, Hawaii, IEEE Press, 2002

Examples


data(yeast)
# Data pre-processing
yeastF <- filter.NA(yeast)
yeastF <- fill.NA(yeastF)
yeastF <- standardise(yeastF)

#### parameter selection
yeastFR <- randomise(yeastF)
cl <- mfuzz(yeastFR,c=16,m=1.25)
mfuzz.plot(yeastFR,cl=cl,mfrow=c(4,5))

tmp  <- partcoef(yeastFR)
F <- tmp[[1]];F.n <- tmp[[2]];F.min <- tmp[[3]]
F > F.min

[Package Mfuzz version 0.0.2 Index]