standardise(Mfuzz)	R Documentation

Standardization of microarray data for clustering.

Description

Standardisation of the expression values of every gene is performed, so that the average expression value is zero and the standard deviation is one.

Usage

standardise(eset)

Arguments

eset 	 object of the class“exprSet”

Value

The function produces an object of the exprSet class with standardised expression values.

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik)

Examples


# HERE SHOULD BE A SHORT EXAMPLE

[Package Mfuzz version 0.0.1 Index]