fill.NA(Mfuzz)	R Documentation

Replacement of missing values

Description

This function for replacing missing values by the average value of the gene or its neightbours.

Usage

fill.NA(eset,mode="mean",k=10)

Arguments

eset 	 object of the class“exprSet”

mode 	 method for replacement of missing values:“mean”- missing values will be replaced by the mean expression value of the gene,“median”- missing values will be replaced by the median expression value of the gene,“knn”- missing values will be replaced by the averging over the correponding expression values of the k-nearest neighbours,“knnw”-same replacement method as “knn”, but the expression values averaged are weighted by the distance to the correponding neighbour

k 	 Number of neighbours, if one of the knn method for replacement is chosen.

Value

The function produces an object of the exprSet class with missing values replaced.

Note

The replacement methods “knn” and “knnw” are computationally intensive. It may be a good idea to run this methods as a ‘lunchtime-’ or ‘overnight’ job.

Author(s)

Matthias E. Futschik (http://itb.biologie.hu-berlin.de/~futschik) and Lokesh Kumar

Examples


data(yeast) # data set includes 17 measurements
yeastF <- filter.NA(yeast) 
yeastF <- fill.NA(yeastF)

[Package Mfuzz version 0.0.1 Index]